changed font family and moved the img folder
