+++ src/components/ui/Upload.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Upload = () => {
+    return (
+        <div>Upload</div>
+    )
+}
+
+export default Upload
