+++ src/styles/globals.css
@@ -2,7 +2,13 @@
 @tailwind components;
 @tailwind utilities;
 
-body {
-    padding: 0;
-    margin: 0;
+@layer base {
+    html {
+      font-family: Proxima Nova, system-ui, sans-serif;
+    }
+    
+    body {
+        padding: 0;
+        margin: 0;
+    }
 }
